# [PyTorch][Vulkan] Fix matrix multiplication performance test binary